requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==3.0.7
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-imagekit==4.0.2
importlib-metadata==1.6.1
pilkit==2.0
Pillow==7.1.2
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
